Time: 8 ms (68.86%), Space: 47.8 MB (74.62%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -1,19 +1,19 @@
 class Solution {
     public int minEatingSpeed(int[] piles, int h) {
-        long l=1;
-        long r=(int)1e9;
+        int l=1;
+        int r=(int)1e9;
         int ans=0;
         while(l<=r){
-            long mid=l+(r-l)/2;
+            int mid=l+(r-l)/2;
             long val=0;
             for(int i=0;i<piles.length;i++){
-                val+=((long)piles[i]+mid-1)/(long)mid;
+                val+=((long)piles[i]+mid-1)/mid;
             }
             if(val>h){
                 l=mid+1;
             }
             else if(val<=h){
-                ans=(int)mid;
+                ans=mid;
                 r=mid-1;
             }
         }
